Add files via upload
In this project, I applied my Data Structure and Algorithm skills, specifically utilizing the Breath-First Search (BFS) method and Queues. The project's objective was to determine the most efficient path to reach the end while minimizing the required space inside the maze. The modules used were Curses, queue and time.